requirements.txt
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
accelerate>=0.25.0
tensorboard>=2.13.0
numpy>=1.24.0
pandas>=1.5.0
scikit-learn>=1.0.0
tqdm>=4.66.0
peft>=0.6.0
nltk>=3.8.0
matplotlib>=3.6.0
seaborn>=0.12.0
requests>=2.28.0
python-dotenv>=1.0.0
rouge-score>=0.1.2
sacrebleu>=2.3.1
bert-score>=0.3.13
language-tool-python>=2.7.1
textstat>=0.7.3
anthropic>=0.5.0
openai>=1.3.0
tiktoken>=0.4.0
jupyter>=1.0.0
notebook>=6.5.0